backend/requirements-lambda.txt
pydantic>=2.12.4,<3.0
boto3>=1.35.0,<2.0
httpx>=0.27.0,<1.0
python-dotenv>=1.0.1,<2.0
python-multipart>=0.0.18,<1.0
pymupdf>=1.25.0,<2.0
